% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFilter_class.R
\name{shinyfilterset}
\alias{shinyfilterset}
\alias{data_filter}
\title{Define a filter set}
\usage{
shinyfilterset(..., id = NULL)

data_filter(id = NULL, column_data, column_name,
  filter_ui = c("picker", "select", "numeric_min", "slider",
  "numeric_max", "numeric_range", "switch"), sort = TRUE,
  options = list())
}
\arguments{
\item{\dots}{The filters, or HTML elements. Define filters with \code{data_filter}, see Examples.}
}
\description{
An R6 class to define a 'filter set', a collection of ui elements (sliders, pickers, etc.) 
to filter a dataset, for use in shiny applications.
}
