% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_list_definition.R
\name{from_list_definition}
\alias{from_list_definition}
\title{Convert a list definition from YAML for shinyfilterset}
\usage{
from_list_definition(lis, ...)
}
\description{
Convert a list definition from YAML for shinyfilterset
}
