% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{numeric_breaks_labels}
\alias{numeric_breaks_labels}
\title{Make a numeric vector into categories based on breaks}
\usage{
numeric_breaks_labels(x, breaks, round_digits = 1)
}
\description{
Cuts a vector based on the breaks argument (plus the min and max values, which should NOT)
be supplied in the breaks argument), and returns a vector of equal length with the category the data fall in.
}
