% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafilter_classes.R
\name{shinyfilterset}
\alias{shinyfilterset}
\alias{data_filter}
\title{Define a filter set}
\usage{
shinyfilterset(..., id = NULL, all_data_on_null = TRUE)

data_filter(
  id = NULL,
  column_data,
  column_name,
  filter_ui = c("picker", "select", "checkboxes", "numeric_min", "slider",
    "numeric_max", "numeric_range", "switch"),
  updates = FALSE,
  sort = TRUE,
  all_choice = NULL,
  search_method = c("equal", "regex"),
  options = list(),
  ui_section = 1
)
}
\arguments{
\item{\dots}{The filters, or HTML elements. Define filters with \code{data_filter}, see Examples.}
}
\description{
An R6 class to define a 'filter set', a collection of ui elements (sliders, pickers, etc.) 
to filter a dataset, for use in shiny applications.
}
\details{
This function makes a shinyfilterset objects, which has a number of methods and properties. 
id, elements, filters
ui(), apply(), reactive(), update()
The shinyfilterset is made up of \code{data_filter} objects, and optionally as many HTML tags as you 
like, for layout purposes.  These have to be constructed with \code{shiny::tags} (see Examples).
Each data filter has properties and methods of its own, most of which don't need to be called directly
by the user.
}
